requirements.txt
# Symphony Core Document Management Workflow - Dependencies
# Python 3.11+ required

# Core dependencies
PyYAML>=6.0.1
click>=8.1.7
pathlib>=1.0.1; python_version < '3.11'

# Markdown parsing (for Sprint 2-3)
markdown-it-py>=3.0.0
mdit-py-plugins>=0.4.0

# Future: API clients (v1.1+)
anthropic>=0.25.0

# Development dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.1
types-PyYAML>=6.0.12

# Testing utilities
faker>=20.1.0
